+++ anyblok/tests/test_field.py
@@ -7,10 +7,10 @@
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
 from anyblok.tests.testcase import TestCase, DBTestCase
-from anyblok.field import Field, FieldException, Function
-from anyblok.column import Integer, String
+from anyblok.field import Field, FieldException, Function, JsonRelated
+from anyblok.column import Integer, String, Json
 from anyblok import Declarations
-from sqlalchemy import func
+from sqlalchemy import func, types
 
 
 Model = Declarations.Model
@@ -148,3 +148,163 @@ class TestField2(DBTestCase):
 
         registry = self.init_registry(add_in_registry)
         registry.Test.val2
+
+    def define_field_json_related(self):
+
+        @register(Model)
+        class Test:
+
+            id = Integer(primary_key=True)
+            properties = Json()
+            name = JsonRelated(json_column='properties', keys=['name'])
+
+    def define_field_json_related2(self):
+
+        @register(Model)
+        class Test:
+
+            id = Integer(primary_key=True)
+            properties = Json()
+            name = JsonRelated(json_column='properties', keys=['sub', 'name'])
+
+    def define_field_json_related_multi_keys(self):
+
+        @register(Model)
+        class Test:
+
+            id = Integer(primary_key=True)
+            properties = Json()
+            name = JsonRelated(json_column='properties', keys=['sub', 'name'])
+
+    def test_field_json_related_hasattr(self):
+        registry = self.init_registry(self.define_field_json_related)
+        registry.Test.name
+
+    def test_field_json_related_get_1(self):
+        registry = self.init_registry(self.define_field_json_related)
+        t = registry.Test.insert()
+        self.assertIsNone(t.name)
+
+    def test_field_json_related_get_2(self):
+        registry = self.init_registry(self.define_field_json_related)
+        t = registry.Test.insert(properties={})
+        self.assertIsNone(t.name)
+
+    def test_field_json_related_get_3(self):
+        registry = self.init_registry(self.define_field_json_related)
+        t = registry.Test.insert(properties={'name': 'jssuzanne'})
+        self.assertEqual(t.name, 'jssuzanne')
+
+    def test_field_json_related_get_4(self):
+        registry = self.init_registry(self.define_field_json_related2)
+        t = registry.Test.insert(
+            properties={'sub': {'name': 'jssuzanne'}})
+        self.assertEqual(t.name, 'jssuzanne')
+
+    def test_field_json_related_set_1(self):
+        registry = self.init_registry(self.define_field_json_related)
+        t = registry.Test.insert()
+        t.name = 'jssuzanne'
+        self.assertEqual(t.properties, {'name': 'jssuzanne'})
+
+    def test_field_json_related_set_2(self):
+        registry = self.init_registry(self.define_field_json_related)
+        t = registry.Test.insert(properties={})
+        t.name = 'jssuzanne'
+        self.assertEqual(t.properties, {'name': 'jssuzanne'})
+
+    def test_field_json_related_set_3(self):
+        registry = self.init_registry(self.define_field_json_related)
+        t = registry.Test.insert(properties={'name': 'other'})
+        t.name = 'jssuzanne'
+        self.assertEqual(t.properties, {'name': 'jssuzanne'})
+
+    def test_field_json_related_set_4(self):
+        registry = self.init_registry(self.define_field_json_related2)
+        t = registry.Test.insert(properties={'sub': {'name': 'other'}})
+        t.name = 'jssuzanne'
+        self.assertEqual(t.properties, {'sub': {'name': 'jssuzanne'}})
+
+    def test_field_json_related_exp_1(self):
+        registry = self.init_registry(self.define_field_json_related)
+        Test = registry.Test
+        Test.insert()
+        query = registry.Test.query().filter(
+            Test.name.cast(types.String) == '"jssuzanne"')
+        self.assertFalse(query.count())
+
+    def test_field_json_related_exp_2(self):
+        registry = self.init_registry(self.define_field_json_related)
+        Test = registry.Test
+        Test.insert(properties={})
+        query = registry.Test.query().filter(
+            Test.name.cast(types.String) == '"jssuzanne"')
+        self.assertFalse(query.count())
+
+    def test_field_json_related_exp_3(self):
+        registry = self.init_registry(self.define_field_json_related)
+        Test = registry.Test
+        Test.insert(properties={'name': 'jssuzanne'})
+        query = registry.Test.query().filter(
+            Test.name.cast(types.String) == '"jssuzanne"')
+        self.assertTrue(query.count())
+
+    def test_field_json_related_exp_4(self):
+        registry = self.init_registry(self.define_field_json_related2)
+        Test = registry.Test
+        Test.insert(properties={'sub': {'name': 'jssuzanne'}})
+        query = registry.Test.query().filter(
+            Test.name.cast(types.String) == '"jssuzanne"')
+        self.assertTrue(query.count())
+
+    def test_field_json_related_with_adapter(self):
+
+        def add_in_registry():
+
+            @register(Model)
+            class Test:
+
+                id = Integer(primary_key=True)
+                properties = Json()
+                name = JsonRelated(json_column='properties', keys=['name'],
+                                   set_adapter=str, get_adapter=int)
+
+        registry = self.init_registry(add_in_registry)
+        t = registry.Test.insert()
+        self.assertIsNone(t.name)
+        t.name = 1
+        self.assertEqual(t.properties, {'name': '1'})
+        t.properties['name'] = '2'
+        self.assertEqual(t.name, 2)
+
+    def test_field_json_related_with_adapter_with_method(self):
+
+        def add_in_registry():
+
+            @register(Model)
+            class Test:
+
+                id = Integer(primary_key=True)
+                properties = Json()
+                name = JsonRelated(json_column='properties', keys=['name'],
+                                   set_adapter='fromint', get_adapter='toint')
+
+                def fromint(self, value):
+                    if value:
+                        return str(value)
+
+                    return value
+
+                def toint(self, value):
+                    if value:
+                        return int(value)
+
+                    return value
+
+        registry = self.init_registry(add_in_registry)
+        t = registry.Test.insert()
+        self.assertIsNone(t.name)
+        t.name = 1
+        self.assertEqual(t.properties, {'name': '1'})
+        t.properties['name'] = '2'
+        self.assertEqual(t.name, 2)